+++ runtime/rust/python-wheel/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "triton-distributed"
+name = "triton-distributed-rs"
 version = "0.1.1"
 description = "Distributed LLM Framework"
 # readme = "README.md"
@@ -13,11 +13,11 @@ dependencies = [
 ]
 
 # [project.scripts]
-# triton-distributed = "triton_distributed:main"
+# triton-distributed = "triton_distributed_rs:main"
 
 [tool.maturin]
-module-name = "triton_distributed._core"
-python-packages = ["triton_distributed"]
+module-name = "triton_distributed_rs._core"
+python-packages = ["triton_distributed_rs"]
 python-source = "python"
 
 [build-system]
@@ -69,10 +69,6 @@ indent-width = 4
 # --explicit-package-bases: WAR errors about duplicate module names used
 #   throughout project such as launch_workers.py
 # explicit_package_bases = true
-
-# NOTE
-# We ignore model.py explcitly here to avoid mypy errors with duplicate modules
-exclude = ["model.py"]
 
 # --ignore-missing-imports: WAR too many errors when developing outside
 #   of container environment with PYTHONPATH set and packages installed.
