+++ runtime/rust/python-wheel/python/triton_distributed_rs/__init__.py
@@ -3,8 +3,7 @@ from functools import wraps
 from typing import Any, AsyncGenerator, Callable, Type
 
 from pydantic import BaseModel, ValidationError
-
-from triton_distributed._core import DistributedRuntime
+from triton_distributed_rs._core import DistributedRuntime
 
 
 def triton_worker():
